wallet/requirements.txt
flask
jsonify
datetime
boto3
uuid
PyJWT